# MPI datatype trans optimization