--- a/vlasovsolver/cpu_trans_map_amr.hpp
+++ b/vlasovsolver/cpu_trans_map_amr.hpp
@@ -190,6 +190,7 @@ void update_remote_mapping_contribution_amr(dccrg::Dccrg<spatial_cell::SpatialCe
 
 void flagSpatialCellsForAmrCommunication(const dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid,
                                          const std::vector<CellID>& localPropagatedCells,
+                                         const std::vector<CellID>& remoteTargetCells,
                                          const uint dimension,
                                          const bool reset);