@markusbattarbee 2021-02-25
And with `--diff` given to `vslvdiff_DP`:
```

Statistics on file: size 9000 min = 1.00011e+06 max = 2.27098e+06 average = 1.7351e+06 standard deviation 3091.99
Statistics on file: size 9000 min = 1.00011e+06 max = 2.27098e+06 average = 1.7351e+06 standard deviation 3091.99
The absolute 0-distance between both datasets is 209.829
The relative 0-distance between both datasets is 9.23959e-05
The average-shifted absolute 0-distance between both datasets is 211.274
The average-shifted relative 0-distance between both datasets is 9.30323e-05
The absolute 1-distance between both datasets is 345875
The relative 1-distance between both datasets is 2.21489e-05
The average-shifted absolute 1-distance between both datasets is 345580
The average-shifted relative 1-distance between both datasets is 2.213e-05
The absolute 2-distance between both datasets is 4771.79
The relative 2-distance between both datasets is 2.85836e-05
The average-shifted absolute 2-distance between both datasets is 4769.82
The average-shifted relative 2-distance between both datasets is 2.85718e-05
```