@markusbattarbee 2021-02-25
I'll also run a more standard testpackage test