@ursg 2021-02-25
Indeed, shouldn't this end up having the same data where it is needed, and only have potential differences due to summation ordering?