@markusbattarbee 2021-02-25
This was 843 timesteps instead of the 100 used in Flowthrough_amr so it's possible it is just that, or it's possible there's a bug somewhere. I'll run Flowthrough_amr and post some diff pictures.