@markusbattarbee 2021-02-25
Well, something is amiss :)
```
vlsvdiff_DP --diff Flowthrough_amr_dev/bulk.0000001.vlsv Flowthrough_amr_MPItrans/bulk.0000001.vlsv proton/vg_rho 0
INFO Reading in two files.
Statistics on file: size 4664 min = 1.04839e+06 max = 2.15588e+06 average = 1.50635e+06 standard deviation 6801.63
Statistics on file: size 4664 min = 1.04839e+06 max = 2.15897e+06 average = 1.50638e+06 standard deviation 6801.22
The absolute 0-distance between both datasets is 9766.62
The relative 0-distance between both datasets is 0.00453022
The average-shifted absolute 0-distance between both datasets is 9803.78
The average-shifted relative 0-distance between both datasets is 0.00454745
The absolute 1-distance between both datasets is 304585
The relative 1-distance between both datasets is 4.33536e-05
The average-shifted absolute 1-distance between both datasets is 463004
The average-shifted relative 1-distance between both datasets is 6.59024e-05
The absolute 2-distance between both datasets is 32922.3
The relative 2-distance between both datasets is 0.000305823
The average-shifted absolute 2-distance between both datasets is 32824.4
The average-shifted relative 2-distance between both datasets is 0.000304913
```
diff proton/vg_rho 0 in y-slice:
![AMR_y_map_DIFF_proton_vg_rho_0000001](https://user-images.githubusercontent.com/11755746/109149176-e14e1c80-776f-11eb-8e7e-cc58d48ba696.png)


ranks:
![AMR_y_map_vg_rank_0000001](https://user-images.githubusercontent.com/11755746/109149129-d2676a00-776f-11eb-984a-0297a0afa56d.png)