@markusbattarbee 2021-02-25
With just two tasks it looks even weirder:
diff proton/vg_rho 0:
![AMR1_y_map_DIFF_proton_vg_rho_0000001](https://user-images.githubusercontent.com/11755746/109152943-d77ae800-7774-11eb-89ff-fdb9d9d5618a.png)

ranks:
![AMR_y_map_vg_rank_0000001](https://user-images.githubusercontent.com/11755746/109152897-c6ca7200-7774-11eb-9db8-bfdfbf62e932.png)