@markusbattarbee 2021-02-25
Some plots to verify the flags:
dark and pale regions are different ranks (2 processes)
purple means the flag is active, check the pseudocolor description to verify which direction is plotted
So in the relevant direction, there ordering must be:
pale grey cells <-> two pale purple cells  <-> two dark purple cells  <-> dark grey cells

These all look correct to me.

<img width="814" alt="Screenshot 2021-02-25 at 16 41 34" src="https://user-images.githubusercontent.com/11755746/109169695-c2a84f80-7788-11eb-8c51-77a3050de5b9.png">
<img width="806" alt="Screenshot 2021-02-25 at 16 41 24" src="https://user-images.githubusercontent.com/11755746/109169702-c50aa980-7788-11eb-8bb9-c6f5db99a7c5.png">
<img width="810" alt="Screenshot 2021-02-25 at 16 41 07" src="https://user-images.githubusercontent.com/11755746/109169705-c63bd680-7788-11eb-8fe7-7a3cf8c92831.png">
<img width="808" alt="Screenshot 2021-02-25 at 16 40 57" src="https://user-images.githubusercontent.com/11755746/109169713-c6d46d00-7788-11eb-99fe-01cd93ef4177.png">