@markusbattarbee 2021-02-25
Hmpf. I would like to understand why this isn't working as intended, but at least with relatively small vspaces the time spent in transfer-stencil-data-x is only of the order of 2-5 percent of translate, so this won't cause a huge increase even if it worked. Still would be nice to understand this better.