@markusbattarbee 2021-02-25
Verified that after 1 step only, diffs are still on a numerical level. At subsequent steps, they slowly grow (Still below 1e-7 at 5 steps). Turning FS and ACC off did not fix things.