@markusbattarbee 2021-03-01
So I tried adding the
`updateRemoteVelocityBlockLists(mpiGrid,popID,VLASOV_SOLVER_Y_NEIGHBORHOOD_ID);`
commands to dev and testing that, but after 99 timesteps, diffs are still big - so it's apparently not that dev would require that update.